Create FunctionMeta to store metadata for functions.
This allows us to stop holding state information on functions themselves
(aka stop treating functions as a "bag of properties").

Once this PR lands we will still be storing these function meta objects
on the function itself (in either a `WeakMap` when supported or on a
`__ember_function_meta__` property when `WeakMap` is not present) just
like we store normal object meta.